Data has been recorded from source-websites in JSON form with js
For further alignment and compatibility with the project, java based httpparser code has been deleted and the code has been converted into javascript.

Also for compatibility with the database, in merging phase of the project, the data has been saved in JSON Object arrays for both earthquake and fire instances.